--- src/app/components/Navbar/i18n.js
@@ -1,7 +1,0 @@
-import i18next from 'i18next';
-
-i18next.addResources('es', 'Navbar', {
-  profile: 'Mi Perfil',
-  home: 'Mi Recorrido',
-  getOut: 'Salir'
-});
