--- src/app/components/Navbar/styles.module.scss
@@ -1,84 +1,0 @@
-/* stylelint-disable */
-
-@import 'src/scss/variables/colors';
-@import 'src/scss/variables/variables';
-
-.navbar {
-  background-color: $white;
-  box-shadow: 1px 2px 3px rgba($cod-gray, 0.2);
-  left: 0;
-  padding: 15px 20px;
-  position: fixed;
-  right: 0;
-  top: 0;
-  z-index: 1;
-}
-
-.link {
-  cursor: pointer;
-  transition: all $transition-duration $transition-function;
-
-  &.active,
-  &:hover {
-    color: $limeade;
-
-    .icon path,
-    .icon ellipse {
-      fill: $limeade;
-    }
-  }
-
-  & > i {
-    font-size: 18px;
-  }
-}
-
-.icon {
-  height: 20px;
-  width: 20px;
-}
-
-@media only screen and (max-width: $small-device) {
-  .navbar {
-    bottom: 0;
-    box-shadow: none;
-    padding: 0;
-    top: auto;
-  }
-
-  .title,
-  .logout {
-    display: none;
-  }
-
-  .container-links {
-    flex: 1;
-    justify-content: space-between;
-    padding: 10px 0;
-  }
-
-  .icon {
-    height: 25px;
-    margin-bottom: 5px;
-    margin-right: 0;
-    width: 25px;
-  }
-
-  .link {
-    flex: 1;
-    flex-direction: column;
-    font-size: 14px;
-    justify-content: center;
-    margin-right: 0;
-    padding: 5px 10px;
-
-    &:first-child {
-      border-right: 2px solid $wild-sand;
-    }
-
-    &:hover,
-    &.active {
-      color: $limeade;
-    }
-  }
-}
