+++ src/app/screens/Dashboard/screens/Home/styles.module.scss
@@ -1,25 +1,22 @@
-/* stylelint-disable */
 @import 'src/scss/variables/colors';
 @import 'src/scss/variables/variables';
 
 .container {
-  background-color: $white;
-  border-radius: 10px;
-  bottom: 10px;
-  box-shadow: 0 -2px 10px hsla(211, 100%, 30%, 0.2), -1px 2px 10px hsla(211, 100%, 30%, 0.15);
-  left: 0;
-  margin: 0 auto;
+  margin: 20px auto;
   max-width: 100%;
-  padding: 10px 15px;
-  position: absolute;
-  right: 0;
-  width: 400px;
+  width: 600px;
 }
 
-.container-hidden {
+.title {
+  background-color: $white;
+  box-shadow: 0 1px 2px rgba($cod-gray, 0.2);
+  padding: 15px 10px;
+}
 
-  @extend .container;
-  display: none;
+.chart {
+  height: 260px;
+  max-width: 100%;
+  width: 400px;
 }
 
 .modal-img {
@@ -28,21 +25,6 @@
 
 @media only screen and (max-width: $small-device) {
   .container {
-    border-radius: 20px 20px 0 0;
-    bottom: 40px;
-
-    &.open {
-      transform: translateY(-80px);
-    }
-  }
-
-  .btn-inside {
-    display: none;
-  }
-
-  .container-hidden {
-    bottom: 70px;
-    display: flex;
-    top: auto;
+    margin: 0 auto;
   }
 }
