--- src/app/screens/Dashboard/screens/Profile/i18n.js
@@ -1,5 +1,0 @@
-import i18next from 'i18next';
-
-i18next.addResources('es', 'Profile', {
-  textExplanation: 'Porcentaje de tiempo de exposición a la calidad del aire:'
-});
