--- src/app/screens/Dashboard/screens/Profile/styles.module.scss
@@ -1,33 +1,0 @@
-/* stylelint-disable */
-@import 'src/scss/variables/colors';
-@import 'src/scss/variables/variables';
-
-.container {
-  margin: 20px auto;
-  max-width: 100%;
-  width: 600px;
-}
-
-.title {
-  background-color: $white;
-  box-shadow: 0 1px 2px rgba($cod-gray, 0.2);
-  padding: 15px 10px;
-}
-
-.cloud {
-  filter: drop-shadow(2px 2px 3px rgba($cod-gray, 0.2));
-  height: 50px;
-  width: 50px;
-}
-
-.chart {
-  height: 260px;
-  max-width: 100%;
-  width: 400px;
-}
-
-@media only screen and (max-width: $small-device) {
-  .container {
-    margin: 0 auto;
-  }
-}
