+++ src/app/screens/Dashboard/screens/Trip/components/GoogleMap/index.js
