+++ src/app/screens/Dashboard/screens/Trip/components/Notification/index.js
