+++ src/app/screens/Dashboard/screens/Trip/components/Notification/styles.module.scss
