+++ src/app/screens/Dashboard/screens/Trip/i18n.js
@@ -1,0 +1,5 @@
+import i18next from 'i18next';
+
+i18next.addResources('es', 'Trip', {
+  finish: 'Terminar recorrido'
+});
