+++ src/app/screens/Dashboard/screens/Trip/styles.module.scss
@@ -1,0 +1,32 @@
+/* stylelint-disable */
+@import 'src/scss/variables/colors';
+@import 'src/scss/variables/variables';
+
+.container {
+  background-color: $white;
+  border-radius: 10px;
+  bottom: 10px;
+  box-shadow: 0 -2px 10px hsla(211, 100%, 30%, 0.2), -1px 2px 10px hsla(211, 100%, 30%, 0.15);
+  left: 0;
+  margin: 0 auto;
+  max-width: 100%;
+  padding: 10px 15px;
+  position: absolute;
+  right: 0;
+  width: 400px;
+}
+
+@media only screen and (max-width: $small-device) {
+  .container {
+    border-radius: 20px 20px 0 0;
+    bottom: 40px;
+
+    &.open {
+      transform: translateY(-80px);
+    }
+  }
+
+  .btn-inside {
+    display: none;
+  }
+}
