+++ src/scss/application.scss
@@ -21,7 +21,8 @@ body {
   padding: 0;
 }
 
-button {
+button,
+.button {
   border-radius: $default-border-radius;
   color: $cod-gray;
   cursor: pointer;
@@ -66,17 +67,9 @@ button {
   margin: 0 auto;
   max-width: 100%;
   min-height: 100vh;
-  padding-top: 70px;
   width: $content-max-width;
 }
 
 .full-height {
   height: 100%;
-}
-
-@media only screen and (max-width: $small-device) {
-  .main-content {
-    padding-bottom: 50px;
-    padding-top: 0;
-  }
 }
